Add configuration option for timeout on retry
When a secondary objective other than the minimization of stalls
is registered, SLOTHY minimizes stalls first, and then does another
optimization pass for the secondary objective, fixing the minimum
number of stalls.

This commit adds a configuration option to register a separate timeout
for the optimization of this secondary objective.