AArch64: Minor cleanup of instruction constructors
It is good style to initialize in `__init__()` all attributes
that an object may work with. That was previously not the case
for `Instruction`, which e.g. used the `datatype` attribute only
if the instruction assembly had a datatype annotation. Now, this
and other similar attributes are initialized to `None` in `__init__()`.