Fix inconsistency in visualization style of optimization
With `config.visualize_expected_performance` set, a nice graph with legend
would be shown, but not so if `visualize_expected_performance == False`.

This commit largely copy-pastes the visualization code from
`visualize_expected_performance` to the other visualization
functions, so that a legend is printed in any case.

Ultimately, this should be cleaned up a bit and a helper function
introduced for the pretty printing of a performance/reordering graph.