Print solver's bound for # of cycles in generated assembly
The solver provides us with a lower/upper bound for the objective
it is minimizing/maximizing.

In the case where we minimize the number of cycles, print the
objective bound in the emitted assembly, so it's clear if the
code is optimal according to the model, or can perhaps be improved
by running SLOTHY with larger timeout.