Allow setting `functional_only = False`
`config.constraints.functional_only` is a wrapper property around
`model_latencies` and `model_functional_units`. However, previously
it would only be supported to set `functional_only = True`, in which
case both `model_latencies` and `model_functional_units` would be
set to `False` -- attempting `functional_only = False`, in turn,
would be silently ignored, which is certainly not desired.

This commit modifies the `functional_only` setter to always set
both `model_functional_units` and `model_latencies` to the negation
of the value that's being specified.