Add 'altnerating' strategy for naive interleaving
When `split_heuristic_preprocess_naive_interleaving` is enabled,
SLOTHY preprocesses the input by naively reordering instructions
according to their depths in the computational flow graph.

This commit introduces another naive interleaving strategy
"alternate" which will make SLOTHY alternate evenly between
instructions tagged with `interleaving_class=0/1`. This is
useful when two sequential blocks of code are to be interleaved
as evenly as possible, which is common in scalar/Neon hybrids.