Make address offset fixup configurable
Address offset fixup is a feature which leverages commutativity
relations such as

```
   ldr X, [A], #immA;
   str Y, [A, #immB]
==
   str Y, [A, #(immB+immA)]
   ldr X, [A], #immA
```

to achieve greater instruction scheduling flexibility in SLOTHY.

When one enables this feature, one MUST ensure that registers which are
used for addresses are not used in any other instruction than load and
stores. Otherwise, the use of this feature is unsound (one may see ldr/
str instructions with increment reordered with instructions depending
on the address register).

Originally, address offset fixup was on by default and not configurable,
which is dangerous, esp. if SLOTHY does not yet detect violations of the
above condition.

As a step towards fixing this, this commit makes address offset
configurable for the AArch64 architecture model.

The Armv8.1-M architecture model is not compatible with the new option:
For the time being, address offset fixup must remain on in this case.