feat: implement <w>-specifier to indicate a dedicated 32-bit version for some instructions
The <w>-specifier allows having 32-bit and 64-bit versions of instructions in one class
which saves some work. While parsing, mnemonics are compared first to find the matching class
for an instruction. Since add<w> would not be identified as an add-instruction, we first must replace
<w> (if present) by the corresponding regex pattern and call re.match(srcline, pattern) afterward. A generic solution
for parsing specifiers which modify the mnemonic could be a future task.