move export of CP-SAT model
Right now we are exporting the CP-SAT model after _solve()
has returned.
However, at the end of _solve(), we add the found solution as hints
to the model such that sub-sequent solving of the same model
(with a different objective) run faster.
This resulted in the exported model not being the one that was
being solved originally.

This commit moves the export.