AArch64: Organize shift-by-immediate Neon instructions
This commit adds the parent class

- VShiftImmediateBasic

for the instructions `ushr` and `shl`, and

- VShiftImmediateRounding

for the instruction `srshr`.

It also adds a new instruction `sshr` to the `VShiftImmediateBasic`
category, and `urshr` to `VShiftImmediateComplex`.

The motivation for those categories is that they feature in Arm's SWOGs:
Most CPUs seem to have the same instruction characteristics for all of
them, so having a parent class helps modelling.