Armv7-M: Fix fusion when registers in ldm/ldrd overlap
Previously ldm and ldrd fusion would break if
the same register is used as address as one of
the outputs (and it's not the last output).
This commit fixes that by changing the fusion to
re-order the ldr overwriting the address to the
very end in case there is an overlap.

Note that this is not needed for stm/strd as
there you cannot have an overlap.

Additionally, it removes unnecessary restrictions
disallowing Rd=Ra for ldrb/ldrh/ldr.