refactor: Major example refactoring
This commit changes the directory structure of the examples directory. It splits the examples by arch and contains one folder for each class of examples (e.g., Kyber-related asm files). In addition, these subfolders also contain the python scripts implementing the optimization for these asm files. example.py on the top-level is thus merely a wrapper for these.

Moreover, the test-cases that check arch-specific operations, e.g., related to parsing have been removed from the examples dir and moved to tests. They can be run using the test.py file.

The class for running an example or a test now resides in common/OptimizationRunner.py.

To add an example/test, perform the following steps:
1. Create a subdir for the arch if not already present inside examples/tests.
2. For examples: Create a subdir for the group of examples. For tests, this is not used yet but could be.
3. Place asm files in the resp. naive directory.
4. Place _example.py/_test.py in the resp. naive directory. See existing files for examples.
5. In top-level example.py/test.py import the list of examples/tests provided by your new script and add it to the list containing all examples/tests.