Allow ignoring unused-output error for specific instructions
By default, if SLOTHY finds an instruction none of whose outputs
are used by other instructions, nor declared as outputs, it will
fail. This can be globally disabled via the `allow_useless_instructions`
configuration option.

This commit refines this configuration option into a local tag
`ignore_useless_output` which can be added to instructions for
which an unused output should be ignored.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>