AArch64: Don't enforce whitespacing before/after separating tokens
Previously, a whitespace before/after a separating token such
as ',' inside an instruction pattern would be interpreted as
mandatory, leading to parsing failures for valid instructions
strings such as 'ldr q0,[x0]'.

This commit fixes this by allowing any amount of whitespaces before/after
tokens such as ',', '.', '[', ']', regardless of the number of spaces
in the instruction pattern.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>