Refactor Armv8.1-M model to follow instruction template patterns from other models
Our first model was the Armv8.1-M model and it contained a lot of ad-hoc
parsing code. When writing the Neon model, we came up with a much cleaner
way of describing instructions and abstracting away most of the parsing code.
This commit aligns the old Armv8.1-M model with the new way of describing
instructions.

Co-authored-by: willieyz <willie.zhao@chelpis.com>