Armv7-M: Fix ldrd_imm
ldrd_imm uses a pre_index. However, it contains no write method that updates
the immediate in case the address offset fixup modified the pre_index.
This commit adds the missing write function.

This likely does not have any impact on the examples we have, as for
the Cortex-M7, we commonly used the splitting to split an ldrd into
two ldrs.