Add slothy-cli to pypi package
This commit adds the slothy-cli to be be available in $PATH if SLOTHY installed
via pip.
First the slothy-cli script is moved to slothy/cli such that it is included
in the pypi packaging. The root slothy-cli becomes a simple wrapper around it.
Next, the slothy-cli script is registered in the pyproject.toml to be a
callable script.

This was surprisingly simple.