Fix loop detection (disallow cross-arch loop matching)
When parsing a loop, SLOTHY tries to match any of the available loop types
and takes the first that matches.
This could lead to the unfortunately situation that a Armv7E-M loop type
matches for AArch64 code.
For example, previously this code would match a Armv7E-M BranchLoop:

count .req x2

mov count, #16

start:
    add x5, x5, x4
    add x7, x5, x1
    ldr x5, [x0, #4]
    add x5, x5, x7
    subs count, count, #2
    b.ne start

This makes absolutely no sense and would ultimately result in a mysterious
error message that b.ne is not a known instruction.

This commit fixes that by only trying to parse loop types for the current
architecture.