AArch64: Add d-form ldr
This commit adds support for d-form ldrs in three forms:
ldr <Da>, [<Xc>]
ldr <Da>, [<Xc>], <imm>
ldr <Da>, [<Xc>, <imm>]!

The Cortex-A55, Cortex-A72, and Neoverse N1 models are extended accordingly.
For Cortex-A55, there is a considerable difference compared to the q-form
equivalent, and, we hence closely follow the SWOG.

For Cortex-A72 and Neoverse N1, there seems to be no difference in the
performance characteristic between q-form and d-form.
We follow the existing approximations for q-form ldrs.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>