Neon: Set addr for 3 instructions
For the selftest to properly detect which registers are used as addresses,
the model needs to set the addr property. This was missing for 3 instructions:

1) ld2 { <Va>.<dt>, <Vb>.<dt> }[<index>], [<Xa>], <imm>"
2) ldr <Wd>, [<Xa>, <Wb>, SXTW <imm>]
3) ld1r {<Va>.<dt>}, [<Xa>], <imm>

This commit fixes that and also adds a unit test for all memory operations
(some currently can't be tested as they are not supported by the A55 model).

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>