AArch64: Do not write literals with #
SLOTHY accepts literals as #=1234 and =1234 just like GAS, and unlike LLVM
which only accepts the proper =1234.
However, SLOTHY before incorrectly always outputs #=1234 which breaks if you
want to assembly it using LLVM (or if you are in a multi-step optimization
and run SLOTHY on the output again).
This commit fixes SLOTHY by separating the parsing and writing of immediates
and literals.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>